Separated kernel selector from editor header
